Move Cfg class to chipset.py
Delete file
Move class to chipset.py
Remove imports from common.py
