--- chipsec/cfg/common.py
@@ -1,31 +1,0 @@
-#!/usr/bin/python
-#CHIPSEC: Platform Security Assessment Framework
-#Copyright (c) 2010-2020, Intel Corporation
-#
-#This program is free software; you can redistribute it and/or
-#modify it under the terms of the GNU General Public License
-#as published by the Free Software Foundation; Version 2.
-#
-#This program is distributed in the hope that it will be useful,
-#but WITHOUT ANY WARRANTY; without even the implied warranty of
-#MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-#GNU General Public License for more details.
-#
-#You should have received a copy of the GNU General Public License
-#along with this program; if not, write to the Free Software
-#Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
-#
-#Contact information:
-#chipsec@intel.com
-#
-
-class Cfg:
-    def __init__(self):
-        self.CONFIG_PCI    = {}
-        self.REGISTERS     = {}
-        self.MMIO_BARS     = {}
-        self.IO_BARS       = {}
-        self.MEMORY_RANGES = {}
-        self.CONTROLS      = {}
-        self.BUS           = {}
-        self.XML_CONFIG_LOADED = False
